pyproject.toml
[tool.poetry]
name = "delta-context"
version = "0.1.0"
description = ""
authors = ["grahlnn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
whisperx = { git = "https://github.com/m-bain/whisperx.git" }
transformers = "^4.36.2"
accelerate = "^0.26.1"
datasets = { extras = ["audio"], version = "^2.16.1" }
setuptools = "^69.0.3"
spacy = { extras = ["cuda11x"], version = "^3.7.2" }
audio-separator = "^0.13.0"
yt-dlp = "^2023.12.30"
openai = "^1.7.1"
pydub = "^0.25.1"
bilibili-api-python = "^16.1.1"
optimum = "^1.16.1"
inflect = "^7.0.0"
toml = "^0.10.2"
selenium = "^4.16.0"
torch = { version = "^2.1.2", source = "pytorch" }
torchvision = { version = "^0.16.2", source = "pytorch" }
torchaudio = { version = "^2.1.2", source = "pytorch" }
mdx23 = { git = "https://github.com/GrahLnn/MVSEP-MDX23-Colab_v2.git" }
tiktoken = "^0.5.2"
moviepy = "^1.0.3"
nltk = "^3.8.1"
ctranslate2 = "^3.24.0"
faster-whisper = "^0.10.0"
pyannote-audio = "^3.1.1"
voicefixer = "^0.1.3"
# resemble-enhance = { git = "https://github.com/GrahLnn/resemble-enhance.git" }
librosa = "^0.10.1"
siamese = {git = "https://github.com/GrahLnn/Siamese-pytorch.git"}
openai-whisper = "^20231117"
whisper-timestamped = "^1.14.4"
redlines = "^0.4.2"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
